Add files via upload
So first we would like to point to the sensor. we are using the IR sensor here to detect the presence of any person in front of the sensor. like if anyone comes in front of the sensor the sensor will detect and pass this information. 
Now the next part is the controller or the CPU of this project. We are using Arduino Uno as the Controller of this project. Of course, we need a controller which will control all the things in this project. like getting input processing and giving output. It is the brain part of this system. It is getting input from the IR sensor. Basically, we interface the IR sensor with Arduino. It is also known as the Arduino counter. Arduino gets the input from the IR sensor and processes the input and according to the input the Arduino will take action according to the condition.
The 16×2 LCD display which displays all the detail about the project and the counting of the people. people went outside the room and inside the room also the total number of people inside the room. Liquid crystal display using the analog signal to lit the crystal for the desired numbers or letters.
 The working of bidirectional counter project there are two sensors named the first sensor is S1 and the other sensor is s2 When a person crosses the S1 sensor first the system will wait for the s2 sensor to be crossed. if both conditions will true the project will count the person who went inside the sensor. next, if the s2 sensor is crossed first and then the s1 sensor will be crossed by a person then it will count the person who goes outside the system or room.